moved customViews folder tried to fix circle card view radius to always be circular
