+++ app/src/main/java/com/sealstudios/pokemonApp/ui/customViews/fabFilter/CircleCardView.kt
@@ -1,8 +1,7 @@
-package com.sealstudios.pokemonApp.ui.util.customViews.fabFilter
+package com.sealstudios.pokemonApp.ui.customViews.fabFilter
 
 import android.content.Context
 import android.util.AttributeSet
-import android.util.Log
 import android.view.WindowInsets
 import com.google.android.material.card.MaterialCardView
 import kotlin.math.min
@@ -15,10 +14,10 @@ class CircleCardView @JvmOverloads constructor(
 ) : MaterialCardView(context, attrs, defStyleAttr) {
 
     override fun setRadius(radius: Float) {
-        super.setRadius(
-            if (radius > height / 2 || radius > width / 2) min(height, width) / 2f
-            else radius
-        )
+
+        val r = if (radius > height / 2 || radius > width / 2) min(height, width) / 2f
+        else radius
+        super.setRadius(r)
     }
 
     override fun onApplyWindowInsets(insets: WindowInsets?): WindowInsets? {
