+++ app/src/main/res/xml/splash_screen_layout_scene.xml
@@ -1,70 +1,34 @@
 <?xml version="1.0" encoding="utf-8"?>
 <MotionScene xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:motion="http://schemas.android.com/apk/res-auto"
-    xmlns:app="urn:oasis:names:tc:xliff:document:1.2">
+    xmlns:motion="http://schemas.android.com/apk/res-auto">
 
     <Transition
-        motion:constraintSetStart="@id/pokeball_thrown"
-        motion:constraintSetEnd="@+id/pokeball_landed"
         motion:autoTransition="animateToEnd"
+        motion:constraintSetEnd="@+id/pokeball_landed"
+        motion:constraintSetStart="@id/pokeball_thrown"
         motion:duration="1000">
-<!--       <KeyFrameSet>-->
-<!--           <KeyAttribute-->
-<!--               motion:motionTarget="@+id/pokeball_icon"-->
-<!--               motion:framePosition="0"-->
-<!--               android:scaleX="1" />-->
-<!--           <KeyAttribute-->
-<!--               motion:motionTarget="@+id/pokeball_icon"-->
-<!--               motion:framePosition="0"-->
-<!--               android:scaleY="1" />-->
-<!--           <KeyAttribute-->
-<!--               motion:motionTarget="@+id/pokeball_icon"-->
-<!--               motion:framePosition="40"-->
-<!--               android:translationY="-50dp" />-->
-<!--           <KeyAttribute-->
-<!--               motion:motionTarget="@+id/pokeball_icon"-->
-<!--               motion:framePosition="70"-->
-<!--               android:translationY="10dp" />-->
-<!--           <KeyAttribute-->
-<!--               motion:motionTarget="@+id/pokeball_icon"-->
-<!--               motion:framePosition="100"-->
-<!--               android:scaleY="2" />-->
-<!--           <KeyAttribute-->
-<!--               motion:motionTarget="@+id/pokeball_icon"-->
-<!--               motion:framePosition="100"-->
-<!--               android:translationY="150dp" />-->
-<!--           <KeyAttribute-->
-<!--               motion:motionTarget="@+id/pokeball_icon"-->
-<!--               motion:framePosition="100"-->
-<!--               android:rotation="540" />-->
-<!--           <KeyAttribute-->
-<!--               motion:motionTarget="@+id/pokeball_icon"-->
-<!--               motion:framePosition="100"-->
-<!--               android:scaleX="2" />-->
-<!--       </KeyFrameSet>-->
-        <KeyFrameSet >
+        <KeyFrameSet>
             <KeyAttribute
-                motion:motionTarget="@id/pokeball_icon"
+                android:translationY="-100dp"
                 motion:framePosition="20"
-                android:translationY="-100dp" />
+                motion:motionTarget="@id/pokeball_icon" />
         </KeyFrameSet>
     </Transition>
 
-    <ConstraintSet android:id="@+id/pokeball_thrown">
-    </ConstraintSet>
+    <ConstraintSet android:id="@+id/pokeball_thrown"></ConstraintSet>
 
     <ConstraintSet android:id="@+id/pokeball_landed">
         <Constraint
+            android:id="@+id/pokeball_icon"
+            android:layout_width="100dp"
+            android:layout_height="100dp"
+            android:rotation="720"
             android:scaleX="2"
             android:scaleY="2"
-            android:rotation="720"
-            motion:layout_constraintEnd_toEndOf="parent"
-            android:layout_height="100dp"
-            android:layout_width="100dp"
             motion:layout_constraintBottom_toBottomOf="parent"
-            motion:layout_constraintVertical_bias="0.9"
-            motion:layout_constraintTop_toTopOf="parent"
+            motion:layout_constraintEnd_toEndOf="parent"
             motion:layout_constraintStart_toStartOf="parent"
-            android:id="@+id/pokeball_icon" />
+            motion:layout_constraintTop_toTopOf="parent"
+            motion:layout_constraintVertical_bias="0.9" />
     </ConstraintSet>
 </MotionScene>
